删除用户统一

+++ src/views/sysmanage/userManage/index.vue
@@ -484,23 +484,28 @@ export default {
       this.editUserForm = {}
     },
     showDeleteDialog(username, id) {
-      this.deleteUserDialogVisible = true
       this.deleteUserName = username
       this.deleteUserId = id
+      this.deleteAUser()
     },
-
     deleteAUser() {
       const ids = []
       ids.push(this.deleteUserId)
-      deleteUser(ids).then(response => {
-        if (response.code !== 0) {
-          return this.$message.error('删除用户失败,请稍后再试')
-        }
-        this.deleteUserDialogVisible = false
-        this.deleteUserId = 0
-        this.deleteUserName = ''
-        this.getUserList()
-        this.$message.success('删除用户信息')
+      this.$confirm('此操作将永久删除该数据, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        deleteUser(ids).then(response => {
+          if (response.code !== 0) {
+            return this.$message.error('删除用户失败,请稍后再试')
+          }
+          this.deleteUserDialogVisible = false
+          this.deleteUserId = 0
+          this.deleteUserName = ''
+          this.getUserList()
+          this.$message.success('删除用户信息')
+        })
       })
     },
